web-assistant/src/requirements.txt

python-dotenv
ibm_watson
ibm_cloud_sdk_core
bs4